Add missing param 1 config option to Jasco 35931
Parameter 1 controls the power-on behavior of the light (i.e. after a power outage). Right now the available options both involve the light coming on after power is restored, but there is another option that restores the light's previous state (which means if it was off before losing power, it will stay off after power is restored).

I found the 3rd option listed here: http://manuals-backend.z-wave.info/make.php?lang=en&sku=39723&cert=ZC10-17075703, but was unable to find any other detailed documentation for this device. However, I own multiple of these devices and have already tested this change on my local zwave-js instance.

This is an important option IMO - I've woken up in the middle of the night because my bedroom lights came on after a power fluctuation 😆