Ecolink TBZ500 Z-Wave Smart Thermostat
This is my first PR, so please be gentle. I've created this Config file for this new device. It is an RCS-based device used by both Ecolink and Nortek. I found a GoControl device (gocontrol_gc-tbz48.json) that contained many similar parameters, and built on that to build this file based upon the details in the Ecolink manual. I used Lint in the GitPod editor with no errors. This is about as far as I can take it, if someone can take over, verify and merge.