Additional Product ID for HS-WX300
I read in the changelog that the HomeSeer HS-WX300 had added support, and I got some new dimmer/switch combos for finishing the basement.

When I added it (via zwavejs2mqtt), it very much didn't look like it was supported. A little digging and it seemed my device had a different ID code. Not 100% sure this is how you add support for this, but thought I'd try to help.