Update parameter 15
I have 3 of these switches (I don't have any of the dimmer type). The motion timeout setting has not been working, finally realized this should be a valueSize of 2. Additionally it accepts values outside of the options, so I've removed the limit on manual entry and added a description. Tested and working well for my 3 of these devices.