feat(config): let device config override interview values (#4588)
Some devices report their capabilities incorrectly in ways that break
the normal interview process, such as by advertising an incorrectly
formatted bitmap of thermostat setpoint types.

This change adds a new compat flag "commandClasses.interview" that lets
a device configuration file skip the interview process for a given
command class and endpoint and use the provided values instead. It should
be usable with any application command class.

NOTE: This change is intended as an proof of concept for further discussion.