feat(config): save battery device life by seven percent
On change from 1000 to 500: I do not think shorter sleep is worth a user parameter that then has to be checked. Active battery time is 12.5mA; sleep 1uA, so 1/2 second savings is 26 extra daya a year (700 device). More for 500 devices.

By the time this is called all pending messages are cleared, so all the work is done.