feat(config): add NEO Cool Cam Repeater
Merging should close #5348 and #5824. Although the numbers and model are different the manual for both is the same.

There is one issue that should be tested by someone with the device before merging.  The manual description for the Association groups is not logical. First the referenced param numbers appear low by one based on context.  Second, it appears all groups are for low alarms despite the parameter label.