feat(config): add PM-B400ZW-N
First extract attempt from ZWA to close #5849 (help wanted issue). Tried to pick up the bot import #5871, but had unexpected message (separate repository?), so started from fresh.
Only real area of concern for me is param 1. It appears from the manual that once a value over 0 is added, the device enables the 4th byte (0x01). I do not see how to do that separately with masking. It seems the function of the parameter is to turn off the switch if the value falls below the setting value.
Other than that I used the 'label' to explain the enable_disable for params 3 &4 rather than add new labels with options to simplify a bit.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>