feat(config): add Heatit Water Stopper
This will close #6533 Notes: This is a clone of the  Custos (mfg 0x0270), except the mfg info was changed to 0x019b.  I copied the Custos file, then to speed up the review ;-) modified the file to use some templates.  It is still mildly verbose, but down to 297 lines from 577.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>