fix: allow State as a valid HA option for 6 States with No Data options
Not being a HA user and a beginner in ZWave JS, this is highly likely not correct, but I can handle that.

This is a potential hack to address #6584.  For the 6 states that allow the "No Data" option as state parameters, the device mfr might not provide a state parameter (optional per the ZW spec) and in this case ZWave JS sends the state, not the state parameter.  This throws an HA validation error.  The intent of this PR is to add the "state" to the "state parameter" list so that validation will succeed. Conveniently all the "states" are higher than any Zwave "state parameter" options, so the device will never send that "state parameter".  It is only to inform HA that the value is allowed and means no data was provided.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>