feat(config): modify ZEN30 to match Zooz change log
Had to assume minimum level for this device was 1.5 (I'm assuming that 1.05 gets read as 1.5 per normal convention) The current file seemed to make that assumption too as param 22 was missing, 24, 25. and 26 were already added. I did not understand some of the "If" firmware statements, so changed them to how I understand it.

Quite a mess of firmware revisions. Params 28, 29 and 30 were added for 1.11, missing from 2.0, added back in 2.11, dropped for 3.0 and finally added for good in 3.10.  Anyway take a look if you have anytime for configs.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>