fix(config): blacklist window covering command class for Vesternet VES-ZW-MOT-018
Newer versions of the VES-ZW-MOT-018 module purport to support the Window Covering Command Class but this seems to cause issues between Home Assistant and Z-Wave JS (presumably Home Assistant tries to use the new command class which is failing in some way at the device level).

Blacklisting the command class seems to revert the behaviour to how the older version of the module works, returning functionality to Home Assistant.