chore: release v12.12.0
We were informed by Silicon Labs that 700/800 series controllers have a hardware watchdog that can reset the controller when it becomes unresponsive. This feature is now enabled by default in Z-Wave JS and should prevent the controller from hanging indefinitely.

In case this causes new issues, the feature can be disabled by setting the environment variable `ZWAVEJS_DISABLE_WATCHDOG` to any non-empty value, or by setting the driver option `features.watchdog` to `false`.

Please let us know if this actually helps or not.

### Features
* Enable hardware watchdog on 700/800 series controllers (#6954)
* Add method to query supported RF regions (#6957)
* Add notification variable for Door/Window tilt state (#6958)

### Bugfixes
* Fixed an issue where value metadata for unknown notification events with known notification types would only be created if the CC version was exactly 2 (#6959)